Removed file_fastq status patching from submit.py
status field has permission restricted_fields, so when the change was made
to use a users credentials for patching and posting, the file_fastq item creation
was failing for non-admin users. This removes the patching of status on file_fastq
so the items will be created with default status ('uploading'). If an upload has failed,
users will still be able to reupload the file without the status change.