Fix excel and non-ascii gene list bugs
Specifically, ensure an excel submission without a title is not given
the title "None" and catch any gene inputs containing non-ascii
characters to give error and provide user feedback.

Note: non-ascii characters in the title remain acceptable, but the
non-ascii characters in gene names will throw an error when performing
the get request.