Restrict gene list overwrite to institution
Previously, gene lists and associated documents were overwritten if a
user of the same project submitted a gene list of the same title.
However, this caused permission errors if the user institution differed
from the existing gene list's institution.

Now, to overwrite a gene list, the user's project AND institution must
match that of the gene list. This solves the permissions issue, but once
gene lists are seen in a project-specific manner, users will not be able
to overwrite visible gene lists from different institutions.